requirements.txt
asgiref==3.5.2
async-generator==1.10
attrs==21.4.0
certifi==2022.5.18
cffi==1.15.0
cryptography==37.0.2
Django==4.0.4
djangorestframework==3.13.1
factory-boy==3.2.1
Faker==13.11.1
h11==0.13.0
idna==3.3
outcome==1.1.0
pycparser==2.21
pyOpenSSL==22.0.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.1
selenium==4.1.5
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
trio==0.20.0
trio-websocket==0.9.2
urllib3==1.26.9
wsproto==1.1.0
